feat: 11th project

+++ src/utils/Api.js
@@ -19,7 +19,7 @@ class Api {
       return this._processServerResponse(res);
     });
   }
-  addNewCard({ name, link }) {
+  addNewCard(name, link) {
     return fetch(`${this._link}cards`, {
       headers: this._headers,
       method: "POST",
@@ -43,13 +43,13 @@ class Api {
       return this._processServerResponse(res);
     });
   }
-  sendUserData(profileData) {
+  sendUserData(userName, userAbout) {
     return fetch(`${this._link}users/me`, {
       headers: this._headers,
       method: "PATCH",
       body: JSON.stringify({
-        name: profileData.username,
-        about: profileData.description,
+        name: userName,
+        about: userAbout,
       }),
     }).then((res) => {
       return this._processServerResponse(res);
@@ -64,21 +64,22 @@ class Api {
       return this._processServerResponse(res);
     });
   }
-  putCardLike(cardId) {
-    return fetch(`${this._link}cards/${cardId}/likes`, {
-      headers: this._headers,
-      method: "PUT",
-    }).then((res) => {
-      return this._processServerResponse(res);
-    });
-  }
-  deleteCardLike(cardId) {
-    return fetch(`${this._link}cards/${cardId}/likes`, {
-      headers: this._headers,
-      method: "DELETE",
-    }).then((res) => {
-      return this._processServerResponse(res);
-    });
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._link}cards/${cardId}/likes`, {
+        headers: this._headers,
+        method: "PUT",
+      }).then((res) => {
+        return this._processServerResponse(res);
+      });
+    } else {
+      return fetch(`${this._link}cards/${cardId}/likes`, {
+        headers: this._headers,
+        method: "DELETE",
+      }).then((res) => {
+        return this._processServerResponse(res);
+      });
+    }
   }
 }
